Add DynUNet and Static UNet
